+++ app/src/main/java/bachelor/test/locationapp/presenter/positioning/LocationData.kt
@@ -1,6 +1,6 @@
 package bachelor.test.locationapp.presenter.positioning
 
-data class LocationData(val xPos: Float = 1f, val yPos: Float = 2.7f, val zPos: Float = 1f, val qualityFactor: Int = -1) {
+data class LocationData(val xPos: Double = 1.0, val yPos: Double = 2.7, val zPos: Double = 1.0, val qualityFactor: Int = -1) {
 
     override fun toString(): String {
         return "$xPos, $yPos, $zPos, $qualityFactor"
